Fix crop_size type in face detection
Corrected the crop_size type in the detect_faces method.

+++ data_pre-processing/preprocess_your_images.py
@@ -48,7 +48,7 @@ for root, dirs, files in tqdm.tqdm(os.walk(img_root_dir)):
                 p = os.path.join(root, name)
                 img = cv2.imread(p)
                 # if img.shape[0]>256 and img.shape[1]
-                faces = mtcnn.detect_faces(Image.fromarray(img[:, :, ::-1]), min_face_size=64, crop_size=(128, 128))
+                faces = mtcnn.detect_faces(Image.fromarray(img[:, :, ::-1]), min_face_size=64, crop_size=(128, 128).astype('float32'))
                 if len(faces) == 0:
                     print("skip")
                     continue
